fix: fix and update texture model with tf 2.14
- replaced instance normalization with groupnorm (n=-1)
- replaced checkpoint loading with keras load model

fixes:
Issue where the texture model was working but outputting garbage. This was due to it loading empty checkpoints.
This was not caught because expect_partial() on .restore that allowed one to restore a blank model.
Originally this issue was due to git stash pop from an outdated texture setup (brfi3983 theory)